% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influential_measures_script.R
\name{influential}
\alias{influential}
\title{Influential Measures Calculation}
\usage{
influential(model, type, threshold = NULL)
}
\arguments{
\item{model}{lm model object}

\item{type}{name of influence measure}

\item{threshold}{optional threshold value}
}
\value{
list contains influence scores and influential points, plot visualizing influential points
}
\description{
Influential Measures Calculation
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
output1 <- influential(model, type = "cooks", threshold = 0.3)
output2 <- influential(model, type = "dffits", threshold = 0.5)
output3 <- influential(model, type = "hadi", threshold = 0.5)
}
