% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influential_measures_script.R
\name{influential}
\alias{influential}
\title{Influential Measures Calculation}
\usage{
influential(model, type, threshold = NULL)
}
\arguments{
\item{model}{lm model object}

\item{type}{name of influence measure}

\item{threshold}{optional threshold value}
}
\value{
list contains influence scores and influential points, plot visualizing influential points
}
\description{
The Influential Points Analysis package includes effective methods for identifying and evaluating significant data points in linear regression models. This program helps users in identifying data points with a significant effect on the regression model's parameters by computing several influence measures such as Cook's Distance, DFFITS, and Hadi's Influence Measure. The package additionally includes visualization features that allow users to quickly identify influential points using appealing lollipop charts.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
output1 <- influential(model, type = "cooks", threshold = 0.3)
output2 <- influential(model, type = "dffits", threshold = 0.5)
output3 <- influential(model, type = "hadi", threshold = 0.5)
}
